+++ image/scripts/variables.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+###########################################################################
+# A file to keep any bash variables. Used in Makefile and also copied
+# into ideide docker image.
+###########################################################################
+
+export this_image_name="ideide"
+export this_image_tag="1.0.0"